리스트 클릭으로 도서정보 가져오기

+++ SpMVC_ReadBook/src/main/java/com/biz/book/mapper/BookDao.java
@@ -7,4 +7,4 @@ import com.biz.book.model.BookVO;
  * 기본 CRUD 메서드를 
  * 별도로 정의하지 않아도 된다.
  */
-public interface BookDao extends GenericaDao<BookVO, Integer>{}
+public interface BookDao extends GenericaDao<BookVO, Long>{}
